Implement new command to make easier to debug single object.
Console log might be a nightmare to read when debuging specific creature. You have to do with all loaded creature that move, melee, cast...

With this patch you are now able to select one or more object to make everything quiet but the chossen unit.

You can also filter all object type so event related to let say players will only be logged.

As some instruction are added for log and there are many call to log i also added a compilation directive so we can disable all log at once by undefining NDEBUG.
That way Release will not be impacted and can even be a little bit faster than before.

Commands added

> server log guidfilter #guid
> server log typefilter #player, creature, item, gob...
> logtarget (add selected unit to the log list)

These commands should be used with filter commands
> server log filter #filtername off
So filter filename is disabled and you are able to see log.
filtername can be "all" so in one command you can enable/disable all filters.

I think this can help alot db dev and even core dev for debugging.

+++ src/game/Spells/SpellMgr.h
@@ -916,7 +916,7 @@ inline bool IsPositiveEffectTargetMode(const SpellEntry* entry, SpellEffectIndex
 
     // If we ever get to this point, we have unhandled target. Gotta say something about it.
     if (entry->Effect[effIndex])
-        DETAIL_LOG("IsPositiveEffectTargetMode: Spell %u's effect %u has unhandled targets (A:%u B:%u)", entry->Id, effIndex,
+        DETAIL_LOG("IsPositiveEffectTargetMode: Spell %u's effect %u has unhandled targets (A:%u B:%u)", entry->Id, uint32(effIndex),
                    entry->EffectImplicitTargetA[effIndex], entry->EffectImplicitTargetB[effIndex]);
     return true;
 }
